Complete first version of the voting system

+++ src/cli/actions.py
@@ -491,7 +491,10 @@ def answer_notification(ctx: ContextApp) -> None:
         notification_id = notifications_messages[selected_notification]
         notification = notifications[notification_id]
         db = ctx.get_database(notification.db_id)
-        #if db.answer_notification()
+        if db.answer_notification(choice, notification):
+            questionary.print("The vote was cast", style="bold")
+            ctx.delete_notification(notifications_messages[selected_notification])
+        else:
+            questionary.print("There was a problem during the voting process", style="boold fg:red")
 
-        ctx.delete_notification(notifications_messages[selected_notification])
         return
